Fixing Event Sending from Engine
Deleted the applicationlevelevents.py module. Also added () to the end of every event that was sent to the engine, thats why nothing was working in the user interface. Also made three functions, is_sqlite_database, _quietly_execute_statement, and _display_error to make sure that the file is a database and that it is properly being connected to.